backend/requirements.txt
fastapi
uvicorn
SQLAlchemy
python-multipart
psycopg2-binary
python-dotenv
passlib
bcrypt==3.2.0
python-jose[cryptography]
asyncpg
aiosqlite